Create LinearSearch.cpp
This C++ program implements a linear search algorithm to find the position of a given element in an array. The user is prompted to input the number of elements in the array and then enter those elements. Afterward, the program prompts the user to input the element they want to search for. The linear search is then performed, and if the element is found, the program displays its position; otherwise, it indicates that the data is not found. The program provides a basic illustration of the linear search technique for array traversal.